+++ docs/source/_static/examples/python/hamiltonian_constructor.py
@@ -63,10 +63,18 @@ print(hamiltonian.get_summary())
 ################################################################################
 
 ################################################################################
+# start-cell-x2c
+# Build a spin-free one-electron X2C Hamiltonian from all-electron orbitals.
+x2c_constructor = create("hamiltonian_constructor", "qdk_x2c", xuncontract=True)
+x2c_hamiltonian = x2c_constructor.run(orbitals)
+# end-cell-x2c
+################################################################################
+
+################################################################################
 # start-cell-list-implementations
 from qdk_chemistry.algorithms import registry
 
 print(registry.available("hamiltonian_constructor"))
-# ['qdk']
+# ['qdk', 'qdk_cholesky', 'qdk_x2c']
 # end-cell-list-implementations
 ################################################################################